requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.7
Deprecated==1.2.18
distro==1.9.0
fastapi==0.115.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jiter==0.10.0
mypy==1.17.1
mypy_extensions==1.1.0
numpy==2.3.2
openai==1.106.1
packaging==25.0
pathspec==0.12.1
pluggy==1.6.0
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
PyGithub==2.5.0
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.5.0
pytest==8.4.2
python-dotenv==1.1.1
python-gitlab==4.9.0
PyYAML==6.0.2
ratelimit==2.2.1
requests==2.32.3
requests-toolbelt==1.0.0
ruff==0.12.12
setuptools==80.9.0
sniffio==1.3.1
starlette==0.38.6
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.30.6
watchfiles==1.1.0
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.3
